Extract out `protocol.rs`
