+++ iroh-doctor/src/doctor.rs
@@ -24,7 +24,7 @@ use iroh::{
     defaults::DEFAULT_STUN_PORT,
     discovery::{dns::DnsDiscovery, pkarr::PkarrPublisher, ConcurrentDiscovery, Discovery},
     dns::default_resolver,
-    endpoint::{self, Connection, ConnectionTypeStream, RecvStream, RemoteInfo, SendStream},
+    endpoint::{self, Connection, ConnectionTypeStream, RemoteInfo},
     key::{PublicKey, SecretKey},
     metrics::MagicsockMetrics,
     relay::RelayUrl,
@@ -33,35 +33,15 @@ use iroh::{
 use iroh_metrics::core::Core;
 use iroh_net_report as netcheck;
 use portable_atomic::AtomicU64;
-use postcard::experimental::max_size::MaxSize;
 use rand::Rng;
 use ratatui::{prelude::*, widgets::*};
-use serde::{Deserialize, Serialize};
-use tokio::{io::AsyncWriteExt, sync};
 use tokio_util::task::AbortOnDropHandle;
 use tracing::warn;
 
 use crate::{
     config::{iroh_data_root, NodeConfig},
-    progress::ProgressWriter,
+    protocol::{self, GuiExt, ProgressBarExt, TestConfig},
 };
-
-/// Trait for progress bar functionality
-pub trait ProgressBarExt: Clone + Send + Sync + 'static {
-    fn set_message(&self, msg: String);
-    fn set_position(&self, pos: u64);
-    fn set_length(&self, len: u64);
-}
-
-/// Trait for GUI functionality
-pub trait GuiExt {
-    type ProgressBar: ProgressBarExt;
-    fn pb(&self) -> &Self::ProgressBar;
-    fn set_send(&self, bytes: u64, duration: Duration);
-    fn set_recv(&self, bytes: u64, duration: Duration);
-    fn set_echo(&self, bytes: u64, duration: Duration);
-    fn clear(&self);
-}
 
 /// Options for the secret key usage.
 #[derive(Debug, Clone, derive_more::Display)]
@@ -215,107 +195,6 @@ pub enum Commands {
         #[clap(long)]
         file: Option<PathBuf>,
     },
-}
-
-/// Possible streams that can be requested.
-#[derive(Debug, Serialize, Deserialize, MaxSize)]
-enum TestStreamRequest {
-    Echo { bytes: u64 },
-    Drain { bytes: u64 },
-    Send { bytes: u64, block_size: u32 },
-}
-
-/// Configuration for testing.
-#[derive(Debug, Clone, Copy)]
-pub struct TestConfig {
-    size: u64,
-    iterations: Option<u64>,
-}
-
-/// Updates the progress bar.
-fn update_pb(
-    task: &'static str,
-    pb: Option<impl ProgressBarExt + Send + 'static>,
-    total_bytes: u64,
-    mut updates: sync::mpsc::Receiver<u64>,
-) -> tokio::task::JoinHandle<()> {
-    if let Some(pb) = pb {
-        pb.set_message(task.to_string());
-        pb.set_position(0);
-        pb.set_length(total_bytes);
-        tokio::spawn(async move {
-            while let Some(position) = updates.recv().await {
-                pb.set_position(position);
-            }
-        })
-    } else {
-        tokio::spawn(std::future::ready(()))
-    }
-}
-
-/// Handles a test stream request.
-async fn handle_test_request(
-    mut send: SendStream,
-    mut recv: RecvStream,
-    gui: &impl GuiExt,
-) -> anyhow::Result<()> {
-    let mut buf = [0u8; TestStreamRequest::POSTCARD_MAX_SIZE];
-    recv.read_exact(&mut buf).await?;
-    let request: TestStreamRequest = postcard::from_bytes(&buf)?;
-    let pb = Some(gui.pb().clone());
-    match request {
-        TestStreamRequest::Echo { bytes } => {
-            // copy the stream back
-            let (mut send, updates) = ProgressWriter::new(&mut send);
-            let t0 = Instant::now();
-            let progress = update_pb("echo", pb, bytes, updates);
-            tokio::io::copy(&mut recv, &mut send).await?;
-            let elapsed = t0.elapsed();
-            drop(send);
-            progress.await?;
-            gui.set_echo(bytes, elapsed);
-        }
-        TestStreamRequest::Drain { bytes } => {
-            // drain the stream
-            let (mut send, updates) = ProgressWriter::new(tokio::io::sink());
-            let progress = update_pb("recv", pb, bytes, updates);
-            let t0 = Instant::now();
-            tokio::io::copy(&mut recv, &mut send).await?;
-            let elapsed = t0.elapsed();
-            drop(send);
-            progress.await?;
-            gui.set_recv(bytes, elapsed);
-        }
-        TestStreamRequest::Send { bytes, block_size } => {
-            // send the requested number of bytes, in blocks of the requested size
-            let (mut send, updates) = ProgressWriter::new(&mut send);
-            let progress = update_pb("send", pb, bytes, updates);
-            let t0 = Instant::now();
-            send_blocks(&mut send, bytes, block_size).await?;
-            drop(send);
-            let elapsed = t0.elapsed();
-            progress.await?;
-            gui.set_send(bytes, elapsed);
-        }
-    }
-    send.finish()?;
-    Ok(())
-}
-
-/// Sends the requested number of bytes, in blocks of the requested size.
-async fn send_blocks(
-    mut send: impl tokio::io::AsyncWrite + Unpin,
-    total_bytes: u64,
-    block_size: u32,
-) -> anyhow::Result<()> {
-    let buf = vec![0u8; block_size as usize];
-    let mut remaining = total_bytes;
-    while remaining > 0 {
-        let n = remaining.min(block_size as u64);
-        send.write_all(&buf[..n as usize]).await?;
-        remaining -= n;
-    }
-    Ok(())
 }
 
 /// Prints a client report.
@@ -508,126 +387,12 @@ impl GuiExt for Gui {
     }
 }
 
-/// Sends, receives and echoes data in a connection.
-async fn active_side<G: GuiExt>(
-    connection: Connection,
-    config: &TestConfig,
-    gui: Option<&G>,
-) -> anyhow::Result<()> {
-    let n = config.iterations.unwrap_or(u64::MAX);
-    if let Some(gui) = gui {
-        let pb = Some(gui.pb());
-        for _ in 0..n {
-            let d = send_test(&connection, config, pb).await?;
-            gui.set_send(config.size, d);
-            let d = recv_test(&connection, config, pb).await?;
-            gui.set_recv(config.size, d);
-            let d = echo_test(&connection, config, pb).await?;
-            gui.set_echo(config.size, d);
-        }
-    } else {
-        let pb: Option<&G::ProgressBar> = None;
-        for _ in 0..n {
-            let _d = send_test(&connection, config, pb).await?;
-            let _d = recv_test(&connection, config, pb).await?;
-            let _d = echo_test(&connection, config, pb).await?;
-        }
-    }
-    Ok(())
-}
-
-/// Sends a test request in a connection.
-async fn send_test_request(
-    send: &mut SendStream,
-    request: &TestStreamRequest,
-) -> anyhow::Result<()> {
-    let mut buf = [0u8; TestStreamRequest::POSTCARD_MAX_SIZE];
-    postcard::to_slice(&request, &mut buf)?;
-    send.write_all(&buf).await?;
-    Ok(())
-}
-
-/// Echoes test a connection.
-async fn echo_test(
-    connection: &Connection,
-    config: &TestConfig,
-    pb: Option<&impl ProgressBarExt>,
-) -> anyhow::Result<Duration> {
-    let size = config.size;
-    let (mut send, mut recv) = connection.open_bi().await?;
-    send_test_request(&mut send, &TestStreamRequest::Echo { bytes: size }).await?;
-    let (mut sink, updates) = ProgressWriter::new(tokio::io::sink());
-    let copying = tokio::spawn(async move { tokio::io::copy(&mut recv, &mut sink).await });
-    let progress = update_pb("echo", pb.cloned(), size, updates);
-    let t0 = Instant::now();
-    send_blocks(&mut send, size, 1024 * 1024).await?;
-    send.finish()?;
-    let received = copying.await??;
-    anyhow::ensure!(received == size);
-    let duration = t0.elapsed();
-    progress.await?;
-    Ok(duration)
-}
-
-/// Sends test a connection.
-async fn send_test(
-    connection: &Connection,
-    config: &TestConfig,
-    pb: Option<&impl ProgressBarExt>,
-) -> anyhow::Result<Duration> {
-    let size = config.size;
-    let (mut send, mut recv) = connection.open_bi().await?;
-    send_test_request(&mut send, &TestStreamRequest::Drain { bytes: size }).await?;
-    let (mut send_with_progress, updates) = ProgressWriter::new(&mut send);
-    let copying =
-        tokio::spawn(async move { tokio::io::copy(&mut recv, &mut tokio::io::sink()).await });
-    let progress = update_pb("send", pb.cloned(), size, updates);
-    let t0 = Instant::now();
-    send_blocks(&mut send_with_progress, size, 1024 * 1024).await?;
-    drop(send_with_progress);
-    send.finish()?;
-    drop(send);
-    let received = copying.await??;
-    anyhow::ensure!(received == 0);
-    let duration = t0.elapsed();
-    progress.await?;
-    Ok(duration)
-}
-
-/// Receives test a connection.
-async fn recv_test(
-    connection: &Connection,
-    config: &TestConfig,
-    pb: Option<&impl ProgressBarExt>,
-) -> anyhow::Result<Duration> {
-    let size = config.size;
-    let (mut send, mut recv) = connection.open_bi().await?;
-    let t0 = Instant::now();
-    let (mut sink, updates) = ProgressWriter::new(tokio::io::sink());
-    send_test_request(
-        &mut send,
-        &TestStreamRequest::Send {
-            bytes: size,
-            block_size: 1024 * 1024,
-        },
-    )
-    .await?;
-    let copying = tokio::spawn(async move { tokio::io::copy(&mut recv, &mut sink).await });
-    let progress = update_pb("recv", pb.cloned(), size, updates);
-    send.finish()?;
-    let received = copying.await??;
-    anyhow::ensure!(received == size);
-    let duration = t0.elapsed();
-    progress.await?;
-    Ok(duration)
-}
-
 /// Accepts connections and answers requests (echo, drain or send) as passive side.
 async fn passive_side(gui: Gui, connection: Connection) -> anyhow::Result<()> {
     loop {
         match connection.accept_bi().await {
             Ok((send, recv)) => {
-                if let Err(cause) = handle_test_request(send, recv, &gui).await {
+                if let Err(cause) = protocol::handle_test_request(send, recv, &gui).await {
                     eprintln!("Error handling test request {cause}");
                 }
             }
@@ -796,7 +561,7 @@ pub async fn accept(
                         if let Ok(stream) = endpoint.conn_type_stream(remote_peer_id) {
                             log_connection_changes(gui.mp.clone(), remote_peer_id, stream);
                         }
-                        let res = active_side(connection, &config, Some(&gui)).await;
+                        let res = protocol::active_side(connection, &config, Some(&gui)).await;
                         gui.clear();
                         let dt = t0.elapsed().as_secs_f64();
                         if let Err(cause) = res {
@@ -807,7 +572,7 @@ pub async fn accept(
                     } else {
                         let gui: Option<&Gui> = None;
                         // silent
-                        active_side(connection, &config, gui).await.ok();
+                        protocol::active_side(connection, &config, gui).await.ok();
                     }
                 }
                 Err(cause) => {
