+++ iroh-doctor/src/lib.rs
@@ -2,3 +2,4 @@ pub mod config;
 pub mod doctor;
 pub mod logging;
 pub(crate) mod progress;
+pub mod protocol;
